backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7


pytest==7.4.2
pytest-cov==4.1.0
pytest-flask==1.2.0


flake8==6.0.0
black==23.7.0
isort==5.12.0

